+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.1.0] - 2025-09-15
+
+### Added
+
+-   Error codes 46, 47, 48, 49
+-   `getDiscordUserPreset` method
+
 ## [4.0.0] - 2024-03-23
 
 ### Added